deploy: 1c69a46158cac0f10048576bfd6116d6c3b5ffb8

+++ examples/events/events.c
@@ -11,7 +11,7 @@ int main(void) {
         glClear(GL_COLOR_BUFFER_BIT);
 
         RGFW_window_swapBuffers(win);
-		RGFW_window_eventWait(win, (u32)RGFW_eventWaitNext);
+		RGFW_window_eventWait(win, RGFW_eventWaitNext);
         while (RGFW_window_checkEvent(win)) {
             switch (win->event.type) {
                 case RGFW_quit: printf("window closed\n"); break;
